+++ lib/brapi/models/v2/macro_series.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Brapi
+  module Models
+    module V2
+      class MacroSeries < Brapi::Model
+        attribute :slug, type: :string
+        attribute :name, type: :string
+        attribute :description, type: :string
+        attribute :unit, type: :string
+        attribute :frequency, type: :string
+        attribute :category, type: :string
+        attribute :start_date, type: :date
+      end
+    end
+  end
+end