+++ lib/brapi/models/v2/treasury_indicators_response.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Brapi
+  module Models
+    module V2
+      class TreasuryIndicatorsResponse < Brapi::Model
+        attribute :results, type: [Brapi::Models::V2::TreasuryBond]
+        attribute :requested_at, type: :time
+        attribute :took, type: :integer
+      end
+    end
+  end
+end